Restructured project files and began simd rng implementation

+++ include/simdmonte/option/option_european.h
@@ -1,12 +1,12 @@
 #pragma once
-#include "option.h"
+#include "simdmonte/option/option.h"
 #include <vector>
 
 class EuropeanOption : public Option {
 public:
   enum class OptionType { Call, Put };
   OptionType type;
-  EuropeanOption(double K, double T, OptionType type);
-  virtual double payoff(const std::vector<double>& path) const override;
+  EuropeanOption(float K, float T, OptionType type);
+  virtual float payoff(const std::vector<float>& path) const override;
   virtual std::unique_ptr<ISimdHelper> get_simd_helper() const override;
 };
